requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-calculation==1.0.0
django-chartjs==2.3.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-formset==0.13.4
django-formtools==2.3
django-formwizard==1.0
django-geoposition==0.3.0
django-import-export==3.2.0
django-jquery==3.1.0
django-offline==0.2
django-pwa==1.1.0
django-utils-six==2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
